+++ src/nurl.ts
@@ -1,4 +1,5 @@
 import {decode, encode} from './punycode'
+import {mask, MaskOptions, match as matchUrlPattern} from './utils'
 import {
     extractPathKey,
     getDynamicPaths,
@@ -8,8 +9,7 @@ import {
     refineQueryWithPathname,
     convertQueryToArray,
     Query,
-    getPathPriority,
-} from './utils'
+} from './utils/internal'
 
 interface URLOptions
     extends Partial<
@@ -30,14 +30,6 @@ interface URLOptions
     baseUrl?: string
     query?: Query
     basePath?: string
-}
-
-export interface MaskOptions {
-    patterns: string[]
-    sensitiveParams: string[]
-    maskChar?: string
-    maskLength?: number
-    preserveLength?: boolean
 }
 
 export default class NURL implements URL {
@@ -483,57 +475,10 @@ export default class NURL implements URL {
     }
 
     static match(url: string, pattern: string) {
-        if (!NURL.canParse(url) || !NURL.canParse(pattern)) {
-            return null
-        }
-
-        const urlSegments = url.split(/[?#]/)[0]?.split('/').filter(Boolean) || []
-        const patternSegments = pattern.split(/[?#]/)[0]?.split('/').filter(Boolean) || []
-
-        if (urlSegments.length !== patternSegments.length) {
-            return null
-        }
-
-        const params: Record<string, string> = {}
-
-        for (let i = 0; i < patternSegments.length; i++) {
-            const patternSegment = patternSegments[i]
-            const urlSegment = urlSegments[i]
-
-            if (isDynamicPath(patternSegment)) {
-                const pathKey = extractPathKey(patternSegment)
-                params[pathKey] = urlSegment
-            } else if (patternSegment !== urlSegment) {
-                return null
-            }
-        }
-
-        return params
+        return matchUrlPattern(url, pattern)
     }
 
-    static mask(
-        url: string,
-        {patterns, sensitiveParams, maskChar = '*', maskLength = 4, preserveLength = false}: MaskOptions,
-    ) {
-        const sortedPatterns = [...patterns].sort((a, b) => (getPathPriority(b) > getPathPriority(a) ? 1 : -1))
-        for (const pattern of sortedPatterns) {
-            const urlObj = NURL.create(url)
-            const matchedParams = NURL.match(urlObj.pathname, pattern)
-            if (!matchedParams) {
-                continue
-            }
-            sensitiveParams.forEach((sensitiveParam) => {
-                if (sensitiveParam in matchedParams) {
-                    const originalValue = matchedParams[sensitiveParam]
-                    const lengthToMask = preserveLength ? originalValue.length : maskLength
-                    matchedParams[sensitiveParam] = maskChar.repeat(lengthToMask)
-                }
-            })
-
-            urlObj.pathname = refinePathnameWithQuery(pattern, matchedParams)
-            return urlObj.toString()
-        }
-
-        return url
+    static mask(url: string, options: MaskOptions) {
+        return mask(url, options)
     }
 }
